Report insertion
Currently, no controls over conformity of data is applied(coming soon).
Updates have been made to the DAOs in order to make the procedure as
clean as possible.

+++ app/src/main/java/it/fmistri/dontdieplease/db/Entry.java
@@ -15,7 +15,7 @@ import androidx.room.ForeignKey;
 public class Entry {
     @ColumnInfo
     @NonNull
-    public Integer report_id = 0;
+    public Long report_id = 0L;
 
     @ColumnInfo
     @NonNull
